backend_0ld/requirements.txt
# Core dependencies
fastapi
uvicorn
aiohttp
pydantic
python-dotenv

# Database
pymongo
motor

# LLM and AI
openai
langgraph
langchain-core
langchain-community

# Async support
asyncio-mqtt
typing-extensions

# Optional: For enhanced functionality
requests
httpx

# Development and testing
pytest
pytest-asyncio